# src/app/api/openapi.yaml
openapi: 3.0.3
info:
  title: Icarus Game Manager API
  description: |
    API for managing Icarus game servers, mods, and configurations.
    
    This API provides endpoints for:
    - Server management (start/stop/restart)
    - Real-time log monitoring
    - Mod installation and management
    - Server configuration
    - Backup and restore operations
  version: 1.0.0
  contact:
    name: Icarus Game Manager
  license:
    name: MIT
servers:
  - url: /api
    description: Local development server
...
paths:
...
  /backups:
    get:
      tags:
        - Backups
      summary: List available backups
      description: Get a list of all available backup files
      responses:
        '200':
          description: List of backups retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  backups:
                    type: array
                    items:
                      $ref: '#/components/schemas/Backup'
        '500':
          description: Failed to get backups list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Backups
      summary: Create a backup
      description: Create a new backup of the server data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Name for the backup
      responses:
        '200':
          description: Backup created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  backupId:
                    type: string
                    description: ID of the created backup
        '500':
          description: Failed to create backup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /backups/restore:
    post:
      tags:
        - Backups
      summary: Restore from backup
      description: Restore server data from a backup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                backupId:
                  type: string
                  description: ID of the backup to restore
      responses:
        '200':
          description: Backup restored successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
        '500':
          description: Failed to restore backup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /backups/download:
    get:
      tags:
        - Backups
      summary: Download a backup file
      description: Download a specific backup file as a ZIP archive
      parameters:
        - name: id
          in: query
          required: true
          description: ID of the backup to download
          schema:
            type: string
      responses:
        '200':
          description: Backup file download
          content:
            application/zip:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              description: Attachment filename
              schema:
                type: string
            Content-Length:
              description: File size in bytes
              schema:
                type: integer
        '400':
          description: Backup ID is required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Backup file not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Failed to download backup
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...
    Backup:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the backup
        name:
          type: string
          description: Display name of the backup
        createdAt:
          type: string
          format: date-time
          description: Date when the backup was created
        size:
          type: number
          description: Size of the backup in bytes
        type:
          type: string
          enum: [manual, automatic]
          description: Type of backup
      required:
        - id
        - name
        - createdAt
        - size
        - type
...